refactor(backend-game-domain): move card models to a value objects folder
