+++ packages/backend/apps/game/backend-game-domain/src/cards/domain/index.ts
@@ -1,15 +1,15 @@
-import { BaseCard } from './models/BaseCard';
-import { Card } from './models/Card';
-import { CardColor } from './models/CardColor';
-import { CardKind } from './models/CardKind';
-import { ColoredCard } from './models/ColoredCard';
-import { DrawCard } from './models/DrawCard';
-import { NormalCard } from './models/NormalCard';
-import { ReverseCard } from './models/ReverseCard';
-import { SkipCard } from './models/SkipCard';
-import { WildCard } from './models/WildCard';
-import { WildDraw4Card } from './models/WildDraw4Card';
 import { AreCardsEqualsSpec } from './specs/AreCardsEqualsSpec';
+import { BaseCard } from './valueObjects/BaseCard';
+import { Card } from './valueObjects/Card';
+import { CardColor } from './valueObjects/CardColor';
+import { CardKind } from './valueObjects/CardKind';
+import { ColoredCard } from './valueObjects/ColoredCard';
+import { DrawCard } from './valueObjects/DrawCard';
+import { NormalCard } from './valueObjects/NormalCard';
+import { ReverseCard } from './valueObjects/ReverseCard';
+import { SkipCard } from './valueObjects/SkipCard';
+import { WildCard } from './valueObjects/WildCard';
+import { WildDraw4Card } from './valueObjects/WildDraw4Card';
 
 export { AreCardsEqualsSpec, CardColor, CardKind };
