+++ packages/backend/apps/game/backend-game-domain/src/cards/domain/valueObjects/SkipCard.ts
