+++ packages/backend/apps/game/backend-game-domain/src/games/domain/fixtures/ActiveGameFixtures.ts
@@ -1,11 +1,11 @@
 import { CardFixtures } from '../../../cards/domain/fixtures/CardFixtures';
-import { CardColor } from '../../../cards/domain/models/CardColor';
-import { DrawCard } from '../../../cards/domain/models/DrawCard';
-import { NormalCard } from '../../../cards/domain/models/NormalCard';
-import { ReverseCard } from '../../../cards/domain/models/ReverseCard';
-import { SkipCard } from '../../../cards/domain/models/SkipCard';
-import { WildCard } from '../../../cards/domain/models/WildCard';
-import { WildDraw4Card } from '../../../cards/domain/models/WildDraw4Card';
+import { CardColor } from '../../../cards/domain/valueObjects/CardColor';
+import { DrawCard } from '../../../cards/domain/valueObjects/DrawCard';
+import { NormalCard } from '../../../cards/domain/valueObjects/NormalCard';
+import { ReverseCard } from '../../../cards/domain/valueObjects/ReverseCard';
+import { SkipCard } from '../../../cards/domain/valueObjects/SkipCard';
+import { WildCard } from '../../../cards/domain/valueObjects/WildCard';
+import { WildDraw4Card } from '../../../cards/domain/valueObjects/WildDraw4Card';
 import { ActiveGame } from '../models/ActiveGame';
 import { GameDirection } from '../models/GameDirection';
 import { GameStatus } from '../models/GameStatus';
