+++ packages/backend/apps/game/backend-game-domain/src/games/domain/models/ActiveGameSlot.ts
@@ -1,4 +1,4 @@
-import { Card } from '../../../cards/domain/models/Card';
+import { Card } from '../../../cards/domain/valueObjects/Card';
 import { BaseGameSlot } from './BaseGameSlot';
 
 export interface ActiveGameSlot extends BaseGameSlot {
