+++ packages/backend/apps/game/backend-game-domain/src/games/domain/specs/CardCanBePlayedSpec.ts
@@ -1,12 +1,12 @@
 import { AppError, AppErrorKind, Spec } from '@cornie-js/backend-common';
 import { Injectable } from '@nestjs/common';
 
-import { Card } from '../../../cards/domain/models/Card';
-import { CardKind } from '../../../cards/domain/models/CardKind';
-import { DrawCard } from '../../../cards/domain/models/DrawCard';
-import { NormalCard } from '../../../cards/domain/models/NormalCard';
-import { ReverseCard } from '../../../cards/domain/models/ReverseCard';
-import { SkipCard } from '../../../cards/domain/models/SkipCard';
+import { Card } from '../../../cards/domain/valueObjects/Card';
+import { CardKind } from '../../../cards/domain/valueObjects/CardKind';
+import { DrawCard } from '../../../cards/domain/valueObjects/DrawCard';
+import { NormalCard } from '../../../cards/domain/valueObjects/NormalCard';
+import { ReverseCard } from '../../../cards/domain/valueObjects/ReverseCard';
+import { SkipCard } from '../../../cards/domain/valueObjects/SkipCard';
 import { ActiveGame } from '../models/ActiveGame';
 import { ActiveGameSlot } from '../models/ActiveGameSlot';
 import { GameOptions } from '../models/GameOptions';
